requirements.txt
uwsgi
django
django-uwsgi
djangorestframework
